8x8 square lattice tested In this case the vector is 2**64 long so an integer kind=8 for its decimal representation is not enough. The code is now using integer kind=16 for the map and related operators. This however prevents the compilation with ifort (where kind=16 is apparently not supported).

+++ ED_HAMILTONIAN_STORED_HxV.f90
@@ -50,7 +50,7 @@ contains
     !
     integer                                :: isite,jsite,hopndx,distance
     real(8)                                :: n_i,n_j
-    integer(8)                             :: m_8,k1_8,k2_8
+    integer(16)                            :: m_8,k1_8,k2_8
     !
     if(.not.Hstatus)stop "ed_buildH_c ERROR: Hsector NOT set"
     isector=Hsector
